refactor: Rename example integration packages
